requirement.txt
black==19.10b0
flake8==3.8.4
pytest==6.2.2
python-dotenv==0.16.0
pytz==2021.1
rich==9.13.0
shioaji==0.3.2.dev1
dolphindb==1.30.0.1